Removed Commerce bidder
