+++ config/config.go
@@ -1149,10 +1149,6 @@ func SetupViper(v *viper.Viper, filename string) {
 	v.SetDefault("adapters.yieldone.endpoint", "https://y.one.impact-ad.jp/hbs_imp")
 	v.SetDefault("adapters.yssp.disabled", true)
 	v.SetDefault("adapters.zeroclickfraud.endpoint", "http://{{.Host}}/openrtb2?sid={{.SourceId}}")
-	v.SetDefault("adapters.commerce.endpoint", "http://{{.Host}}:8001/TestCommerce")
-	v.SetDefault("adapters.commerce.commerceparams.impurl", "https://{{.Host}}.commerce.io/event-collection/beacon/?action=impression")
-	v.SetDefault("adapters.commerce.commerceparams.clickurl", "https://{{.Host}}.commerce.io/event-collection/beacon/?action=click")
-	v.SetDefault("adapters.commerce.commerceparams.conversionurl", "https://{{.Host}}.commerce.io/event-collection/beacon/conversion")
 	v.SetDefault("adapters.koddi.endpoint","http://{{.Host}}:8001/TestCommerce")
 	v.SetDefault("adapters.koddi.commerceparams.impurl", "https://{{.Host}}.koddi.io/event-collection/beacon/?action=impression")
 	v.SetDefault("adapters.koddi.commerceparams.clickurl", "https://{{.Host}}.koddi.io/event-collection/beacon/?action=click")
