+++ openrtb_ext/request.go
@@ -22,7 +22,7 @@ const GPIDKey = "gpid"
 // TIDKey reserved for Per-Impression Transactions IDs for Multi-Impression Bid Requests.
 const TIDKey = "tid"
 
-const CommerceParamKey = "comparams"
+const CommerceParamKey = "commerce"
 
 // NativeExchangeSpecificLowerBound defines the lower threshold of exchange specific types for native ads. There is no upper bound.
 const NativeExchangeSpecificLowerBound = 500
