--- static/bidder-info/commerce.yaml
@@ -1,10 +1,0 @@
-maintainer:
-  email: commerce@pubmatic.com
-gvlVendorID: 999
-capabilities:
-  app:
-    mediaTypes:
-      - banner
-  site:
-    mediaTypes:
-      - banner
